task_service/src/requirements.txt
altair==5.4.1
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
attrs==24.2.0
bcrypt==4.2.1
blinker==1.9.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
dnspython==2.7.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.5
fastapi-cli==0.0.5
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.27.2
idna==3.10
iniconfig==2.0.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
narwhals==1.14.1
numpy==2.1.3
packaging==24.2
pandas==2.2.3
passlib==1.7.4
pillow==11.0.0
pluggy==1.5.0
protobuf==5.28.3
psycopg==3.2.3
psycopg2-binary==2.9.10
pyarrow==18.0.0
pydantic==2.10.1
pydantic_core==2.27.1
pydeck==0.9.1
Pygments==2.18.0
pytest==8.3.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.17
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rich==13.9.4
rpds-py==0.21.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sqlparse==0.5.2
starlette==0.41.3
streamlit==1.40.1
tenacity==9.0.0
toml==0.10.2
tomli==2.1.0
tornado==6.4.2
typer==0.13.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.1
uvloop==0.21.0
watchdog==6.0.0
watchfiles==1.0.0
websockets==14.1
